+++ ZaraCode/InvestmentSimulator.cs
@@ -17,18 +17,18 @@ namespace ZaraCode
             return lastThursday;
         }
 
-        public float GetFinalCapital(List<DailyStock> dataList, float monthlyInvestment)
+        public double GetFinalCapital(List<DailyStock> dataList, double monthlyInvestment)
         {
-            var realInvestment = monthlyInvestment - monthlyInvestment * 0.02f;
-            var currentCapital = 0f;
-            var cDay = 0f;
+            var realInvestment = monthlyInvestment - monthlyInvestment * 0.02;
+            var currentCapital = 0d;
+            var cDay = 0d;
             var init = false;
             var month = 0;
             var lastDailyStock = new DateTime();
             for (int i = 0; i < dataList.Count; i++)
             {
                 var dailyStock = dataList[i];
-                var dailyFluctuation = (float)Math.Round((dailyStock.CloseDay - dailyStock.OpenDay) / dailyStock.OpenDay, 3); 
+                var dailyFluctuation = Math.Round((dailyStock.CloseDay - dailyStock.OpenDay) / dailyStock.OpenDay, 3); 
                 var lastThursday = GetLastThursday(dailyStock);
                 
 
@@ -37,31 +37,31 @@ namespace ZaraCode
                     if (currentCapital == 0)
                     {
                         currentCapital = realInvestment;
-                        currentCapital += (float)Math.Round(dailyFluctuation * currentCapital, 3);
+                        currentCapital += Math.Round(dailyFluctuation * currentCapital, 3);
                         cDay = dailyStock.CloseDay;
                         month = dailyStock.DateTime.Month;
                         init = true;
                     }
                     else
                     {
-                        var fluctuationAfterMarket = (float)Math.Round((dailyStock.OpenDay - cDay) / cDay, 3);
-                        currentCapital += (float)Math.Round(fluctuationAfterMarket * currentCapital + realInvestment, 3);
-                        currentCapital += (float)Math.Round(dailyFluctuation * currentCapital, 3);
+                        var fluctuationAfterMarket = Math.Round((dailyStock.OpenDay - cDay) / cDay, 3);
+                        currentCapital += Math.Round(fluctuationAfterMarket * currentCapital + realInvestment, 3);
+                        currentCapital += Math.Round(dailyFluctuation * currentCapital, 3);
                         cDay = dailyStock.CloseDay;
                     }
                     
                 }
                 else if (init)
                 {
-                    var fluctuationAfterMarket = (float)Math.Round((dailyStock.OpenDay - cDay) / cDay, 3);
-                    currentCapital += (float)Math.Round(fluctuationAfterMarket * currentCapital, 3);
-                    currentCapital += (float)Math.Round(dailyFluctuation * currentCapital, 3);
+                    var fluctuationAfterMarket = Math.Round((dailyStock.OpenDay - cDay) / cDay, 3);
+                    currentCapital += Math.Round(fluctuationAfterMarket * currentCapital, 3);
+                    currentCapital += Math.Round(dailyFluctuation * currentCapital, 3);
                     cDay = dailyStock.CloseDay;
                 }
 
                 lastDailyStock = dailyStock.DateTime;
             }
-            var finalCapital = (float)Math.Round(currentCapital, 3);
+            var finalCapital = Math.Round(currentCapital, 3);
             return finalCapital;
         }
     }
